feat(sync): add connectivity diagnostics and improve pairing UX
Detect and warn users about connectivity issues that may prevent sync:
- Local device: port listen errors, no LAN address detected
- Peer devices: port unreachable (discovered but can't connect)

Pairing flow improvements:
- Show progress when folder sharing takes longer than 5 seconds
- Treat folder sharing timeout as warning, not fatal error
- Increase HTTP timeout to 60 seconds for slow operations

UI/CLI changes:
- Display connectivity warnings with links to Syncthing firewall docs
- Use accessible anchor tags instead of button+onClick for links
- Show local connectivity issues in CLI sync status output

Other:
- Change SyncProgressInfo.Percent from int to float64 for precision

Documentation:
- Add port numbers note (TCP 22100, UDP 21127 differ from Syncthing defaults)
- Clarify monitoring section in headless guide

+++ internal/syncguest/manager.go
@@ -327,7 +327,7 @@ type Status struct {
 	Running        bool
 	DeviceID       string
 	Syncing        bool
-	Progress       int
+	Progress       float64
 	NeedBytes      int64
 	ConnectedPeers int
 	Peers          []PeerStatus
